+++ ChangeLog.rst
@@ -1,6 +1,18 @@
 Prequ Change Log
 ================
 
+General Changes
+~~~~~~~~~~~~~~~
+
+- Remove dependency on setuptools
+
+- Improve error messages on dependency resolution failures
+
+Bug Fixes
+~~~~~~~~~
+
+- Fix unsafe package reporting in the generated output file
+
 1.4.0
 -----